Run Overview screen added

+++ Assets/Scripts/Player/PlayerData.cs
@@ -19,9 +19,9 @@ public class PlayerData : MonoBehaviour
 
     public int bossCount;
 
-    public int score = 0;
-    public int spaceMoney = 0;
-    public int previousMoney = 0;
+    public float score = 0;
+    public float spaceMoney = 0;
+    public float previousMoney = 0;
     public int kills = 0;
 
     public int cutSceneCounter;
